Removing old layout file
