--- app/src/main/java/zurche/simplecalculator/calculator/theme/Color.kt
@@ -1,10 +1,0 @@
-package zurche.simplecalculator.calculator.theme
-
-import androidx.compose.ui.graphics.Color
-
-val InputGray = Color(0xFF333333)
-val PadTeal = Color(0xFF3b978a)
-val OperatorTeal = Color(0xFF3b978a)
-val DarkOperatorTeal = Color(0xFF2a6b62)
-val NumberTeal = Color(0xFF6ecabd)
-val EqualsTeal = Color(0xFF81d1c5)
