+++ app/src/main/java/zurche/simplecalculator/calculator/ui/theme/Color.kt
@@ -2,10 +2,9 @@ package zurche.simplecalculator.calculator.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
+val InputGray = Color(0xFF333333)
+val PadTeal = Color(0xFF3b978a)
+val OperatorTeal = Color(0xFF3b978a)
+val DarkOperatorTeal = Color(0xFF2a6b62)
+val NumberTeal = Color(0xFF6ecabd)
+val EqualsTeal = Color(0xFF81d1c5)
