+++ app/src/main/java/zurche/simplecalculator/calculator/ui/theme/Type.kt
@@ -2,33 +2,29 @@ package zurche.simplecalculator.calculator.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
+import zurche.simplecalculator.app.R
 
-// Set of Material typography styles to start with
+val MontserratBold = FontFamily(Font(R.font.montserrat_bold))
+val MontserratLight = FontFamily(Font(R.font.montserrat_light))
+
 val Typography = Typography(
-    bodyLarge = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 16.sp,
-        lineHeight = 24.sp,
-        letterSpacing = 0.5.sp
-    )
-    /* Other default text styles to override
-    titleLarge = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 22.sp,
-        lineHeight = 28.sp,
-        letterSpacing = 0.sp
+    displayMedium = TextStyle(
+        fontFamily = MontserratBold,
+        fontWeight = FontWeight.Bold,
+        fontSize = 30.sp
     ),
-    labelSmall = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Medium,
-        fontSize = 11.sp,
-        lineHeight = 16.sp,
-        letterSpacing = 0.5.sp
+    displayLarge = TextStyle(
+        fontFamily = MontserratBold,
+        fontWeight = FontWeight.Bold,
+        fontSize = 90.sp
+    ),
+    bodyMedium = TextStyle(
+        fontFamily = MontserratLight,
+        fontWeight = FontWeight.Thin,
+        fontSize = 20.sp
     )
-    */
 )
